Update to upload image
